requirements.txt
jinja2
unidecode
pyyaml
pyside